+++ jest.setup.js
@@ -1,1 +1,2 @@
-global.AbortSignal = require('./packages/primitives/dist/abort-controller').AbortSignal
+global.AbortSignal =
+  require('./packages/primitives/dist/abort-controller').AbortSignal