+++ packages/vm/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @edge-runtime/vm
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`2b412ff`](https://github.com/vercel/edge-runtime/commit/2b412ff42826b1c2e8373479712045265687fc07)]:
+  - @edge-runtime/primitives@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- BREAKING CHANGE: Drop Node.js 12 ([#191](https://github.com/vercel/edge-runtime/pull/191))
+
+### Patch Changes
+
+- Updated dependencies [[`dfa837e`](https://github.com/vercel/edge-runtime/commit/dfa837ee6e41a9951c4f17bd4a638c9d1882d1e2)]:
+  - @edge-runtime/primitives@2.0.0
+
 ## 1.1.0
 
 ### Patch Changes